Changed data type of variable to long int

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,8 +12,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int bits;
-	unsigned int temp;
+	unsigned long int bits;
+	unsigned long int temp;
 /* variable to store the number of bits flipped  */
 	bits = 0;
 /* Exclusive-OR (XOR) numbers to find amount of different bits */
